Windows Universal Platform basic support (#212)
* Update plugin.xml

Added support for Windows 10.

* Create BluetoothSerial.js

* Update BluetoothSerial.js

* Update BluetoothSerial.js

* Update BluetoothSerial.js

* Update BluetoothSerial.js

* Update bluetoothSerial.js

I have modified the subscribe method: since Windows Universal doesn't have PluginResult objects, the success callback can be executed once. To call the success callback multiple times we have to pass it in the arguments array.
Be carefoul though: this edit needs to be tested ALSO with Windows Phone 8, since the code into BluetoothSerial.cs -> line 107 may not properly work...

* Update plugin.xml

Added support for Windows Universal. 
Accordingly to this blog post: https://blog.vjrantal.net/2015/03/12/building-a-cordova-plugin-including-native-code-for-windows-platform/ Cordova (full) support for Windows Universal Platform started with version 3.8.

* Update bluetoothSerial.js

Added the success callback to the array arguments in the method subscribeRawData. This is mandatory in order to execute the callback everytime it's necessary.

* Update BluetoothSerial.js

Added support for methods subscribe, unsubscribe, subscribeRaw, unsubscribeRaw. With this update the plugin basic functionality should be covered.

* Update README.md